+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Coverage Upload Action Changelog
 
+## 2.1.0 (April 9, 2024)
+
+- Added optional input ignoredUsers. ignoredUsers is a comma separated string of authors which will not run this action. Defaulted to just 'dependabot'
+
 ## 2.0.0 (May 18, 2022)
 
 - Modified action to receive body from the payload